requirements.txt
Django>=5.0
djangorestframework>=3.14
pymysql>=1.1
cryptography>=41.0
requests>=2.30
